basically done, need to test more extensively

+++ TriangleTester/src/Points.java
@@ -1,22 +1,22 @@
 public class Points {
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public Points (int x, int y) {
+	public Points (double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 	public String coordinates() {
